rag-pipeline-design/requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
sentence-transformers>=3.0.0
numpy>=1.26.0
pgvector>=0.3.0
